Force combine_hash function to produce int values
Hash values are required to be ints instead of longs in
order to work correctly with some of the operators.

+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/CombineHashFunction.java
@@ -24,6 +24,7 @@ public final class CombineHashFunction
     @SqlType(StandardTypes.BIGINT)
     public static long getHash(@SqlType(StandardTypes.BIGINT) long previousHashValue, @SqlType(StandardTypes.BIGINT) long value)
     {
-        return 31 * previousHashValue + value;
+        // Hash values are required to be ints
+        return (int) (31 * previousHashValue + value);
     }
 }
